Mostly worked on Delegates

+++ CIS560-Project/CIS560-Project/Models/RaceParticipant.cs
@@ -11,12 +11,12 @@ namespace CIS560_Project.Models
         public int RaceParticipantId { get; }
         public int RaceId { get; }
         public int RunnerId { get; }
-        public double Time { get; }
+        public int Time { get; }
         public double AverageHeartRate { get; set; }
 
         public RaceParticipant() { }
 
-        public RaceParticipant(int raceParticipantId, int raceId, int runnerId, double time=0.0, double averageHeartRate=0.0)
+        public RaceParticipant(int raceParticipantId, int raceId, int runnerId, int time=0, double averageHeartRate=0.0)
         {
             RaceParticipantId = raceParticipantId;
             RaceId = raceId;
